refactoring text component
